% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto_df.R
\name{auto_df}
\alias{auto_df}
\title{generates a dataframe listing all the HLS-files}
\usage{
auto_df(folder_filepath, calculate_non_na_pixels = FALSE)
}
\arguments{
\item{folder_filepath}{\code{character}. The filepath of the HLS files.}

\item{calculate_non_na_pixels}{\code{logical}. If \code{TRUE}, calculates the number of non-NA pixels; otherwise, it does not.}
}
\value{
the dataframe for the HLS files
}
\description{
This function create a dataframe of a folder filled with HLS data. Each row of the dataframe is one file.
The columns consist of: - year, - doy, - band, - filepath, - filename, - satellite type, - non_na_pixels.
All of the function do only work if you did not rename the original data, or the names the package gave the files.
}
