% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto_mask.R
\name{auto_mask}
\alias{auto_mask}
\title{applies masking on the grouped files and saves it in a new folder}
\usage{
auto_mask(
  input_filepath,
  output_filepath,
  filterClouds = FALSE,
  filterAdjacent = FALSE,
  filterCloudshadow = FALSE,
  filterSnowice = FALSE,
  filterWater = FALSE,
  filterAerosol_climatology = FALSE,
  filterAerosol_low = FALSE,
  filterAerosol_moderate = FALSE,
  filterAerosol_high = FALSE
)
}
\arguments{
\item{input_filepath}{\code{character}. The filepath of the grouped HLS files. The files need to be grouped using the auto_group command before.}

\item{output_filepath}{\code{character}. The filepath where the filtered data should be saved.}

\item{filterClouds}{\code{logical}. If \code{TRUE}, the function sets all cloud pixels to \code{NA}.}

\item{filterSnowice}{\code{logical}. If \code{TRUE}, the function sets all snow/ice pixels to \code{NA}.}

\item{filterAerosol_climatology}{\code{logical}. If \code{TRUE}, the function sets all aerosol_climatology pixels to \code{NA}.}

\item{filterAerosol_low}{\code{logical}. If \code{TRUE}, sets all pixels with low aerosol content to \code{NA}.}

\item{filterAerosol_moderate}{\code{logical}. If \code{TRUE}, sets all pixels with moderate aerosol content to \code{NA}.}

\item{filterAerosol_high}{\code{logical}. If \code{TRUE}, sets all pixels with high aerosol content to \code{NA}.}

\item{filerAdjacent}{\code{logical}. If \code{TRUE}, the function sets all pixels adjacent to clouds to \code{NA}.}

\item{filterWaster}{\code{logical}. If \code{TRUE}, the function sets all water pixels to \code{NA}.}
}
\value{
All the files will be filtered and saved in the new folder.
}
\description{
This function stack all bands of HLS data from the same day. Thats very handy because the data you download are one banded
tiffs, merging them based on their filename is labour intense. You can also generate an info_df similar to the
df you would get using the hlsmanager::auto_df() function. Thats helpful if you want to check how many scenes you actually have and
whats their aqcuisition date. If you dont have the same number of bands for each date,
you will get an error message. All of the function do only work if you did not rename the original data, or the names the package gave the files.
}
