% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto_group.R
\name{autogroup}
\alias{autogroup}
\title{groups all the bands based on their data and saves them in a new folder}
\usage{
autogroup(input_filepath, output_filepath, give_df_info = FALSE)
}
\arguments{
\item{input_filepath}{\code{character}. The filepath of the downloaded HLS files.}

\item{output_filepath}{\code{character}. The filepath where the grouped data should be saved.}

\item{give_df_info}{\code{logical}. If \code{TRUE}, the function returns a \code{data.frame} where each file is a row and columns include: bands, filepath, filename, and satellite type.}
}
\value{
all the files will be grouped and saved in the new folder,
}
\description{
This function stack all bands of HLS data from the same day. Thats very handy because the data you download are one banded
tiffs, merging them based on their filename is labour intense. You can also generate an info_df similar to the
df you would get using the hlsmanager::auto_df() function. Thats helpful if you want to check how many scenes you actually have and
whats their aqcuisition date. If you dont have the same number of bands for each date,
you will get an error message. All of the function do only work if you did not rename the original data, or the names the package gave the files.
}
